+++ src/builtins/builtins_request.ml
@@ -129,13 +129,13 @@ let () =
       Lang.string (Request.string_of_log (Request.get_log r)))
 
 let () =
-  Lang.add_builtin "request.ready" ~category:`Liquidsoap
+  Lang.add_builtin "request.resolved" ~category:`Liquidsoap
     ~descr:
-      "Check if a request is ready, i.e. is associated to a valid local file. \
-       Unless the initial URI was such a file, a request has to be resolved \
-       before being ready." [("", Value.t, None, None)] Lang.bool_t (fun p ->
+      "Check if a request is resolved, i.e. is associated to a valid local \
+       file. Unless the initial URI was such a file."
+    [("", Value.t, None, None)] Lang.bool_t (fun p ->
       let e = Value.of_value (List.assoc "" p) in
-      Lang.bool (Request.is_ready e))
+      Lang.bool (Request.resolved e))
 
 let () =
   Lang.add_builtin "request.uri" ~category:`Liquidsoap